refactor(routines): keep the timezone card in the routine editor (HOU-470)
Reverts the move of the timezone picker onto the routines list. Per feedback,
the timezone control reads better as the card-styled field it was inside the
routine editor's "When it runs" section. It still edits the single account-wide
preference (every routine shares one zone) — only its placement is restored.
