+++ app/src/locales/en/routines.json
@@ -5,9 +5,7 @@
     "emptyTitle": "Set it and forget it",
     "emptyDescription": "Routines fire on a schedule and only ping you when something actually needs attention.",
     "newRoutine": "New routine",
-    "descriptionShort": "Recurring tasks that fire on schedule and only ping you when something needs attention.",
-    "timezoneLabel": "Timezone",
-    "timezoneHint": "All your routines run in this timezone."
+    "descriptionShort": "Recurring tasks that fire on schedule and only ping you when something needs attention."
   },
   "editor": {
     "back": "Back to routines",
@@ -31,6 +29,8 @@
     "sectionWhen": "When it runs",
     "sectionBehavior": "Behavior",
     "sectionRecent": "Recent runs",
+    "timezoneLabel": "Timezone",
+    "timezoneHint": "All your routines run in this timezone.",
     "nextRun": "Next run {relative}",
     "schedulePreview": "Schedule preview",
     "schedulePreviewHint": "Pick a valid schedule to see when this routine will fire.",
