+++ app/src/locales/es/routines.json
@@ -5,9 +5,7 @@
     "emptyTitle": "Programa y olvídate",
     "emptyDescription": "Las rutinas se disparan en el horario que elijas y solo te avisan cuando algo de verdad necesita atención.",
     "newRoutine": "Nueva rutina",
-    "descriptionShort": "Tareas recurrentes que corren en un horario y solo te avisan cuando algo necesita tu atención.",
-    "timezoneLabel": "Zona horaria",
-    "timezoneHint": "Todas tus rutinas se ejecutan en esta zona horaria."
+    "descriptionShort": "Tareas recurrentes que corren en un horario y solo te avisan cuando algo necesita tu atención."
   },
   "editor": {
     "back": "Volver a rutinas",
@@ -31,6 +29,8 @@
     "sectionWhen": "Cuándo se ejecuta",
     "sectionBehavior": "Comportamiento",
     "sectionRecent": "Ejecuciones recientes",
+    "timezoneLabel": "Zona horaria",
+    "timezoneHint": "Todas tus rutinas se ejecutan en esta zona horaria.",
     "nextRun": "Próxima ejecución {relative}",
     "schedulePreview": "Vista previa del horario",
     "schedulePreviewHint": "Elige un horario válido para ver cuándo se ejecutará esta rutina.",
