+++ ui/routines/src/routines-grid.tsx
@@ -3,12 +3,7 @@
  *
  * The parent tab already labels this surface "Routines", so this view skips
  * a redundant page header and goes straight to a meta row + the list.
- *
- * Timezone is an account-wide setting (one zone for every routine), so its
- * picker lives HERE on the list — not inside each routine's editor. It sits
- * directly under the "New routine" row, capping the list it governs.
  */
-import { useMemo } from "react"
 import {
   cn,
   EmptyHeader,
@@ -16,7 +11,7 @@ import {
   EmptyDescription,
   Button,
 } from "@houston-ai/core"
-import { Plus, Globe } from "lucide-react"
+import { Plus } from "lucide-react"
 import type { Routine, RoutineRun } from "./types"
 import { RoutineRow } from "./routine-row"
 import {
@@ -30,50 +25,12 @@ import {
   type NextFireLabels,
 } from "./labels"
 
-const COMMON_TIMEZONES = [
-  "UTC",
-  "America/Los_Angeles",
-  "America/Denver",
-  "America/Chicago",
-  "America/New_York",
-  "America/Bogota",
-  "America/Mexico_City",
-  "America/Sao_Paulo",
-  "Europe/London",
-  "Europe/Madrid",
-  "Europe/Berlin",
-  "Europe/Athens",
-  "Africa/Lagos",
-  "Asia/Dubai",
-  "Asia/Kolkata",
-  "Asia/Singapore",
-  "Asia/Tokyo",
-  "Australia/Sydney",
-]
-
-function listTimezones(): string[] {
-  try {
-    const supported = (
-      Intl as { supportedValuesOf?: (k: string) => string[] }
-    ).supportedValuesOf?.("timeZone")
-    if (supported && supported.length) return supported
-  } catch {
-    // fall through
-  }
-  return COMMON_TIMEZONES
-}
-
 export interface RoutinesGridProps {
   routines: Routine[]
   /** Most recent run per routine, keyed by routine ID. */
   lastRuns?: Record<string, RoutineRun>
-  /** The account-wide IANA timezone every routine fires in. */
+  /** Account-default IANA timezone — passed to rows for "next run" preview. */
   accountTimezone: string
-  /**
-   * Persist a new account-wide timezone. Changing it re-times every routine.
-   * Omit it (standalone callers) and the timezone bar is hidden.
-   */
-  onTimezoneChange?: (tz: string) => void
   loading?: boolean
   onSelect: (routineId: string) => void
   onCreate?: () => void
@@ -92,61 +49,10 @@ export interface RoutinesGridProps {
   locale?: string
 }
 
-/**
- * Account-wide timezone control. A compact, globe-prefixed `<select>` plus a
- * one-line hint, so the user reads "this zone applies to every routine below".
- */
-function TimezoneBar({
-  accountTimezone,
-  timezones,
-  onTimezoneChange,
-  label,
-  hint,
-  className,
-}: {
-  accountTimezone: string
-  timezones: string[]
-  onTimezoneChange: (tz: string) => void
-  label: string
-  hint: string
-  className?: string
-}) {
-  return (
-    <div className={cn("flex items-center gap-3", className)}>
-      <div className="relative shrink-0">
-        <Globe
-          className="absolute left-2.5 top-1/2 -translate-y-1/2 size-3.5 text-muted-foreground pointer-events-none"
-          strokeWidth={1.75}
-        />
-        <select
-          value={accountTimezone}
-          onChange={(e) => onTimezoneChange(e.target.value)}
-          aria-label={label}
-          className={cn(
-            "appearance-none cursor-pointer rounded-lg border border-border bg-secondary",
-            "pl-8 pr-3 py-1.5 text-xs text-foreground",
-            "transition-colors focus:outline-none focus:border-foreground/40",
-          )}
-        >
-          {timezones.map((tz) => (
-            <option key={tz} value={tz}>
-              {tz}
-            </option>
-          ))}
-        </select>
-      </div>
-      <span className="text-xs text-muted-foreground/70 truncate min-w-0">
-        {hint}
-      </span>
-    </div>
-  )
-}
-
 export function RoutinesGrid({
   routines,
   lastRuns = {},
   accountTimezone,
-  onTimezoneChange,
   loading,
   onSelect,
   onCreate,
@@ -163,13 +69,6 @@ export function RoutinesGrid({
     if (a.enabled !== b.enabled) return a.enabled ? -1 : 1
     return a.name.localeCompare(b.name)
   })
-
-  // The picker lists every zone with the account zone selected; ensure that
-  // zone is present even if the platform's zone list happens to omit it.
-  const timezones = useMemo(() => {
-    const all = listTimezones()
-    return all.includes(accountTimezone) ? all : [accountTimezone, ...all]
-  }, [accountTimezone])
 
   if (loading && routines.length === 0) {
     return (
@@ -217,18 +116,6 @@ export function RoutinesGrid({
             </Button>
           )}
         </div>
-
-        {/* Account-wide timezone — governs every routine in the list below. */}
-        {onTimezoneChange && (
-          <TimezoneBar
-            accountTimezone={accountTimezone}
-            timezones={timezones}
-            onTimezoneChange={onTimezoneChange}
-            label={l.timezoneLabel}
-            hint={l.timezoneHint}
-            className="mb-4"
-          />
-        )}
 
         {/* List card — gray, divides hold rows */}
         <div
